Surya_Pratap_Dash/requirements.txt
tensorflow
pandas
numpy
scikit-learn
matplotlib
opencv-python
streamlit
streamlit-webrtc
av
Pillow